+++ src/evstation.js
@@ -1,10 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import MapboxGL from 'mapbox-gl';
 import DeckGL from '@deck.gl/react';
 import { InteractiveMap } from 'react-map-gl';
 import { ColumnLayer, IconLayer, PathLayer } from '@deck.gl/layers';
-// import { GridLayer } from '@deck.gl/aggregation-layers';
-// import { ScatterplotLayer } from '@deck.gl/layers';
 import 'react-sliding-pane/dist/react-sliding-pane.css';
 import { LinearInterpolator } from '@deck.gl/core';
 import './evstation.css';
@@ -14,7 +12,6 @@ import LeftPane from './LeftPane';
 import SearchFilterPane from './SearchFilterPane';
 import Tooltip from './ToolTip';
 import ButtonGroup from './ButtonGroup';
-// import { useReducer } from 'react';
 
 MapboxGL.accessToken = process.env.REACT_APP_MAPBOX_TOKEN || '';
 
@@ -40,12 +37,7 @@ export default function Map() {
   const [searchTerm, setSearchTerm] = useState('');
   const [manufacturer, setManufacturer] = useState("");
   const [manufacturers, setManufacturers] = useState([]);
-  const [filteredResults, setFilteredResults] = useState([]);
   const [data, setData] = useState([]);
-  // eslint-disable-next-line
-  // const [radiusData, setRadiusData] = useState(Array.from({ length: data.length }, () => Math.random() * 5 + 2));
-  // eslint-disable-next-line
-  const [isClicked, setIsClicked] = useState(false);
   const [showSearch, setShowSearch] = useState(false);
   const [selectedAddress, setSelectedAddress] = useState(null);
   const [mapViewState, setMapViewState] = useState(INITIAL_VIEW_STATE);
@@ -60,22 +52,18 @@ export default function Map() {
   const [sortOrder, setSortOrder] = useState('asc');
   const [tooltipInfo, setTooltipInfo] = useState(null);
   const [mapStyle] = useState(process.env.REACT_APP_MAPBOX_STYLE_URL || '');
-  // eslint-disable-next-line
-  // const [theme, setTheme] = useState('');
-  // const [color, setColor] = useState('');
   const [theme] = useState('');
   const [color] = useState('');
-  // eslint-disable-next-line
   const [chargerNameSearchTerm, setChargerNameSearchTerm] = useState('');
   const [showAllData, setShowAllData] = useState(false);
   const [elevationFactor, setElevationFactor] = useState(0);
 
-  const validData = data.filter(d => {
-    return d.geometry.coordinates[0] !== "NULL" &&
+  const validData = useMemo(() => data.filter(d => (
+    d.geometry.coordinates[0] !== "NULL" &&
     d.geometry.coordinates[1] !== "NULL" &&
     parseFloat(d.geometry.coordinates[0]) !== 0 &&
     parseFloat(d.geometry.coordinates[1]) !== 0
-  });
+  )), [data]);
 
   let lastDataPoint;
   if (validData.length > 0) {
@@ -84,7 +72,6 @@ export default function Map() {
     );
   } else {
     lastDataPoint = null;
-    // console.error('No data available');
   }
   
   const paths = validData.reduce((acc, curr) => {
@@ -128,7 +115,7 @@ export default function Map() {
       data: [lastDataPoint],
       pickable: true,
       iconAtlas: `${process.env.PUBLIC_URL}/car.png`,
-      iconMevstationing: {
+      iconMapping: {
         marker: {x: 10, y: 150, width: 512, height: 512, mask: false}
       },
       getIcon: d => 'marker',
@@ -155,7 +142,6 @@ export default function Map() {
     if (showAllData) {
       intervalId = setInterval(() => {
         setElevationFactor(prevFactor => {
-          // updateTheme();
           if (prevFactor < 14) {
             return prevFactor + 0.5;
           } else {
@@ -174,10 +160,6 @@ export default function Map() {
     }
   }, [showAllData]);
 
-  // const handleClick = () => {
-  //   setShowAllData(prevState => !prevState);
-  // };
-
   const objectToCSVRow = (dataObject) => {
     let dataArray = [];
     for (let o in dataObject) {
@@ -187,14 +169,6 @@ export default function Map() {
     return dataArray.join(',') + '\r\n';
   }
 
-  // const objectToRow = (dataObject) => {
-  //   let ArrayD = [];
-  //   for (let i in dataObject) {
-  //     let innerValue = dataObject[i] === null ? '' : '"' + dataObject[i].toString() + '"';
-  //     ArrayD.push(innerValue);
-  //   }
-  //   return ArrayD.join(',') + '\r\n';
-  // }
   const propertiesData = data.map(d => d.properties);
   const selectedPropertiesData = propertiesData.filter(item => {
     const regionMatch = selectedFilters.region 
@@ -205,16 +179,15 @@ export default function Map() {
     : true;
     const manufacturerMatch = selectedFilters.manufacturer ? item.mnfacr_name === selectedFilters.manufacturer : true;
     const voltTypeMatch = selectedFilters.voltType ? item.volt_type === selectedFilters.voltType : true;
-    const efficiencyValueMatch = Number(selectedFilters.efficiencyValue) ? Number(item.charging_efficiency) === Number(selectedFilters.efficiencyValue) : true;
+    const efficiencyValueMatch = selectedFilters.efficiencyValue !== '' ? Number(item.charging_efficiency) === Number(selectedFilters.efficiencyValue) : true;
     return regionMatch && manufacturerMatch && voltTypeMatch && efficiencyValueMatch
   }); 
   
   const convertToCSV = (selectedPropertiesData) => {
     let csvContent = '\uFEFF';
-    // to header (key value) column name
-    csvContent += objectToCSVRow(Object.keys(selectedPropertiesData[0])); 
+    csvContent += objectToCSVRow(Object.keys(selectedPropertiesData[0]));
   
-    for (let i = 0; i < propertiesData.length; i++) {
+    for (let i = 0; i < selectedPropertiesData.length; i++) {
       csvContent += objectToCSVRow(selectedPropertiesData[i]);
     }
     return csvContent;
@@ -252,12 +225,6 @@ export default function Map() {
     });
   };
 
-  const efficiencyArray = filteredResults.map(result => parseFloat(result.properties.charging_efficiency));
-  const avgEfficiency = efficiencyArray.reduce((acc, val) => acc + val, 0) / efficiencyArray.length;
-  const minEfficiency = Math.min(...efficiencyArray);
-  const maxEfficiency = Math.max(...efficiencyArray);
-
-  
   const handleZoomIn = () => {
     setZoomButtonDisabled(false);
     const interpolator = new LinearInterpolator({
@@ -299,26 +266,14 @@ export default function Map() {
       transitionInterpolator: interpolator,
     });
   };
-  // isClicked ?  left2pane(setPaneIsOpen) info link
-  // const handleClick = ({ x, y, object }) => {
-  //   if (object && object.Longitude !== "NULL" && object.Latitude !== "NULL") {
-  //   setIsClicked(true);
-  //   setPaneIsOpen(false);
-  //   setRightPaneIsOpen(false);
-  //   setClickedChargerName(object.properties.charger_name);
-  //   setChargingEfficiency(1);
-  //   }
-  // };
-  
+
   const handleClosePane = () => {
-    setIsClicked(false);
     setPaneIsOpen(false);
     setRightPaneIsOpen(false);
   };
 
   const handleHomeClick = () => {
     setPaneIsOpen(false);
-    setIsClicked(false);
     setSelectedAddress(null);
     setSearchTerm('');
     setShowSearch(false);
@@ -329,8 +284,7 @@ export default function Map() {
     setHasZoomedIn(false);
     setManufacturer('');
     setVoltType('');
-    // setSpeed('');
-    setSelectedFilters([]);
+    setSelectedFilters({region: '', manufacturer: '', voltType: '', efficiencyValue: ''});
     setFilterStep(0);
   };
 
@@ -354,36 +308,23 @@ export default function Map() {
     }
   };
 
-  
-
-  // if soc under => scatter change
   useEffect(() => {
     if (manufacturer !== "" && !Object.values(selectedFilters).includes(manufacturer)) {
       setSelectedFilters(prevFilters => ({ ...prevFilters, manufacturer: manufacturer }));
     }
   }, [manufacturer, selectedFilters]);
-  
+
   useEffect(() => {
     if (voltType !== "" && !Object.values(selectedFilters).includes(voltType)) {
       setSelectedFilters(prevFilters => ({ ...prevFilters, voltType: voltType }));
     }
   }, [voltType, selectedFilters]);
 
-  // useEffect  efficiency
   useEffect(() => {
     if (efficiencyValue !== "" && !Object.values(selectedFilters).includes(efficiencyValue)) {
       setSelectedFilters(prevFilters => ({ ...prevFilters, efficiencyValue: efficiencyValue }));
     }
   }, [efficiencyValue, selectedFilters])
-
-  // if curr_speed - avg_speed
-  // const [speed, setSpeed] = useState('');
-  // useEffect(() => {
-  //   if (speed !== "" && !Object.values(selectedFilters).includes(speed)) {
-  //     setSelectedFilters(prevFilters => ({ ...prevFilters, speed: speed}));
-  //   }
-  // }, [speed, selectedFilters])
-
 
   useEffect(() => {
   const handleKeyDown = (event) => {
@@ -411,25 +352,14 @@ useEffect(() => {
   fetch(DATA_URL)
     .then(response => response.json())
     .then(json => {
-      console.log(json); 
-      setData(json.features);
-      // setRadiusData(Array.from({ length: json.features.length }, () => Math.random() * 5 + 3));
+      const features = json.features;
+      setData(features);
+      setManufacturers(Array.from(new Set(features.map(f => f.properties.mnfacr_name))).filter(Boolean));
+      setVoltTypes(Array.from(new Set(features.map(f => f.properties.volt_type))).filter(Boolean));
+      setEfficiencyValues(Array.from(new Set(features.map(f => parseFloat(f.properties.charging_efficiency)))).filter(v => !Number.isNaN(v)));
     })
     .catch(error => console.log(error));
 }, []);
-
-// useEffect(() => {
-//   const interval = setInterval(() => {
-//     setRadiusData(Array.from({ length: data.length }, () => Math.random() * 5 + 5));
-//     if (!isClicked && !selectedAddress) {
-//       setRadiusData(Array.from({ length: data.length }, () => Math.random() * 5 + 3));
-//     }
-//   }, 100); 
-
-//   return () => {
-//     clearInterval(interval);
-//   };
-// }, [isClicked, selectedAddress, data.length]);
 
 useEffect(() => {
   const selectedData = validData.find(d => d.properties.charger_id === selectedAddress);
@@ -438,64 +368,31 @@ useEffect(() => {
     setClickedChargerName(selectedData.properties.charger_name);
     setClickedMnfacrName(selectedData.properties.mnfacr_name);
     setClickedModelName(selectedData.properties.model_name);
-  } 
-}, [selectedAddress, validData]); 
+  }
+}, [selectedAddress, validData]);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      const response = await fetch(`${process.env.REACT_APP_API_BASE_URL || ''}/charger`);
-      const data = await response.json();
-  
-      const uniqueManufacturers = Array.from(new Set(data.feature.map(feature => feature.properties.mnfacr_name))).filter(Boolean);
-      setManufacturers(uniqueManufacturers);
+const filteredResults = useMemo(() => {
+  const regionMatch = (feature) => {
+    if (searchTerm === "") return true;
+    const address = feature.properties.address;
+    if (!address) return false;
+    return searchTerm.includes('/')
+      ? searchTerm.split('/').some(r => address.includes(r))
+      : address.includes(searchTerm);
+  };
+  const manufacturerMatch = (feature) =>
+    manufacturer === '' || feature.properties.mnfacr_name === manufacturer;
+  const voltTypeMatch = (feature) =>
+    voltType === '' || feature.properties.volt_type === voltType;
+  const efficiencyMatch = (feature) =>
+    efficiencyValue === '' || Math.abs(feature.properties.charging_efficiency - Number(efficiencyValue)) <= 0.0001;
+  return data.filter(f => regionMatch(f) && manufacturerMatch(f) && voltTypeMatch(f) && efficiencyMatch(f));
+}, [data, searchTerm, manufacturer, voltType, efficiencyValue]);
 
-      const uniqueVoltTypes = Array.from(new Set(data.features.map(feature => feature.properties.volt_type))).filter(Boolean);
-      setVoltTypes(uniqueVoltTypes);
-
-      const uniqueEfficiencys = Array.from(new Set(data.features.map(feature => parseFloat(feature.properties.charging_efficiency)))).filter(Boolean);
-      setEfficiencyValues(uniqueEfficiencys);
-
-      let filterFunction;
-  
-      if (searchTerm === "") {
-        filterFunction = (feature) => true;
-        
-      
-      } else if (searchTerm === "") {
-        filterFunction = (feature) => {
-          const address = feature.properties.speed;
-          return !address.includes('') && !address.includes('') 
-            && (manufacturer === '' || feature.properties.mnfacr_name === manufacturer) 
-            && (voltType === '' || feature.properties.volt_type === voltType)
-            && (efficiencyValue === '' || Math.abs(feature.properties.charging_efficiency - Number(efficiencyValue)) <= 0.0001)
-
-        };
-      } else if (["", ""].includes(searchTerm)) {
-        filterFunction = (feature) => {
-          const address = feature.properties.speed;
-          return searchTerms.some((term) => address.includes(term) && searchTerm === term) 
-            && (manufacturer === '' || feature.properties.mnfacr_name === manufacturer) 
-            && (voltType === '' || feature.properties.volt_type === voltType)
-            && (efficiencyValue === '' || Math.abs(feature.properties.charging_efficiency - Number(efficiencyValue)) <= 0.0001)
-
-        };
-      } else {
-        filterFunction = (feature) => {
-          return (feature.properties.speed.includes("") || feature.properties.speed.includes("")) 
-            && (manufacturer === '' || feature.properties.mnfacr_name === manufacturer) 
-            && (voltType === '' || feature.properties.volt_type === voltType)
-            && (efficiencyValue === '' || Math.abs(feature.properties.charging_efficiency - Number(efficiencyValue)) <= 0.0001)
-
-        };
-      }
-      
-      const results = data.features.filter(filterFunction);
-      setFilteredResults(results);
-      
-    };
-  
-    fetchData();
-  }, [searchTerm, manufacturer, voltType, efficiencyValue]); 
+const efficiencyArray = filteredResults.map(result => parseFloat(result.properties.charging_efficiency));
+const avgEfficiency = efficiencyArray.reduce((acc, val) => acc + val, 0) / efficiencyArray.length;
+const minEfficiency = Math.min(...efficiencyArray);
+const maxEfficiency = Math.max(...efficiencyArray);
 
 
   return (
@@ -543,7 +440,6 @@ useEffect(() => {
           sortResults={sortResults}
           handleAddressClick={handleAddressClick}
           selectedAddress={selectedAddress}
-          // eslint-disable-next-line no-unused-vars
           hasZoomedIn={hasZoomedIn}
           convertToCSV={convertToCSV}
           downloadCSV={downloadCSV}
@@ -568,8 +464,7 @@ useEffect(() => {
     setChargerNameSearchTerm={setChargerNameSearchTerm}    
     />
       <div className="title-container">
-        <h1 className="title-main" style={{color: color}}>TEST PathLayer</h1>
-        {/* <h1 className="title-sub" style={{color: color}} onClick={toggleMapStyle}>서울, 경기/인천, 제주 테스트</h1> */}
+        <h1 className="title-main" style={{color: color}}>EV Station Dashboard</h1>
       </div>
     </DeckGL>
      </div>